Fix building and parsing

+++ build.sh
@@ -14,4 +14,8 @@ RUSTFLAGS="-C target-cpu=native" cargo build --release
 popd
 
 mkdir -p builds
-cp ../instant-distance/target/release/libinstant_distance.so builds/instant_distance.so.$1
+if [ -f ../instant-distance/target/release/libinstant_distance.so ]; then
+    cp ../instant-distance/target/release/libinstant_distance.so builds/instant_distance.so.$1
+elif [ -f ../instant-distance/target/release/libinstant_distance_py.so ]; then
+    cp ../instant-distance/target/release/libinstant_distance_py.so builds/instant_distance.so.$1
+fi
